# opt: prove partial index implication in scanIndexIter