@mgartner 2020-10-14
Comparison to `master` of the micro benchmarks:

```
name                                      old time/op  new time/op  delta
Phases/kv-read-no-prep/OptBuild           45.1µs ± 7%  37.1µs ± 4%  -17.74%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Normalize           17.6µs ± 6%  15.6µs ± 0%  -11.34%  (p=0.008 n=5+5)
EndToEnd/kv-read-const/EndToEnd            143µs ± 5%   128µs ± 2%  -10.07%  (p=0.008 n=5+5)
Phases/kv-read/ExecBuild                  17.0µs ± 7%  15.6µs ± 2%   -8.61%  (p=0.008 n=5+5)
Phases/tpcc-new-order/Parse               2.81ns ± 3%  2.57ns ± 2%   -8.48%  (p=0.008 n=5+5)
Phases/kv-read-no-prep/Parse              15.3µs ± 8%  14.2µs ± 2%   -7.22%  (p=0.008 n=5+5)
Chain/chain-10/Explore                     975µs ± 4%   921µs ± 3%   -5.56%  (p=0.016 n=5+5)
EndToEnd/tpcc-stock-level/EndToEnd         555µs ± 8%   525µs ± 1%   -5.40%  (p=0.016 n=5+5)
Chain/chain-5/Explore                      336µs ± 2%   332µs ± 0%   -0.99%  (p=0.016 n=5+5)
Phases/kv-read/OptBuild                   8.48µs ± 1%  8.49µs ± 2%     ~     (p=0.690 n=5+5)
Phases/kv-read/Normalize                  8.55µs ± 3%  8.56µs ± 2%     ~     (p=1.000 n=5+5)
Phases/kv-read/Explore                    14.5µs ± 3%  14.3µs ± 1%     ~     (p=0.548 n=5+5)
Phases/kv-read-no-prep/Normalize          44.3µs ± 8%  41.1µs ± 6%     ~     (p=0.056 n=5+5)
Phases/kv-read-no-prep/ExecBuild          55.3µs ± 6%  56.3µs ±14%     ~     (p=1.000 n=5+5)
Phases/kv-read-const/Parse                2.63ns ± 5%  2.63ns ± 4%     ~     (p=1.000 n=5+5)
Phases/kv-read-const/Explore               154ns ± 3%   155ns ± 3%     ~     (p=0.833 n=5+5)
Phases/kv-read-const/ExecBuild             890ns ±10%   868ns ± 1%     ~     (p=0.802 n=5+5)
Phases/tpcc-new-order/OptBuild            16.6µs ± 7%  15.8µs ± 2%     ~     (p=0.690 n=5+5)
Phases/tpcc-new-order/Explore             46.7µs ±10%  42.5µs ± 3%     ~     (p=0.151 n=5+5)
Phases/tpcc-delivery/Parse                2.57ns ± 1%  2.55ns ± 2%     ~     (p=0.357 n=5+5)
Phases/tpcc-delivery/Normalize            17.2µs ± 3%  17.7µs ± 1%     ~     (p=0.095 n=5+5)
Phases/tpcc-delivery/Explore              35.9µs ±11%  35.4µs ±12%     ~     (p=0.690 n=5+5)
Phases/tpcc-delivery/ExecBuild            33.6µs ± 3%  34.7µs ± 8%     ~     (p=0.310 n=5+5)
Phases/tpcc-stock-level/Parse             2.53ns ± 1%  2.52ns ± 1%     ~     (p=0.865 n=5+5)
Phases/tpcc-stock-level/OptBuild          62.6µs ± 3%  63.8µs ± 4%     ~     (p=0.222 n=5+5)
Phases/tpcc-stock-level/Normalize         71.6µs ±11%  64.4µs ± 4%     ~     (p=0.095 n=5+5)
Phases/tpcc-stock-level/Explore            186µs ± 8%   188µs ± 4%     ~     (p=0.421 n=5+5)
Phases/tpcc-stock-level/ExecBuild          192µs ± 4%   195µs ± 2%     ~     (p=0.310 n=5+5)
EndToEnd/kv-read/EndToEnd                  173µs ± 9%   166µs ± 3%     ~     (p=0.310 n=5+5)
EndToEnd/kv-read-no-prep/EndToEnd          274µs ± 5%   258µs ±10%     ~     (p=0.151 n=5+5)
EndToEnd/tpcc-new-order/EndToEnd           233µs ± 7%   225µs ± 2%     ~     (p=0.222 n=5+5)
EndToEnd/tpcc-delivery/EndToEnd            207µs ± 8%   197µs ± 2%     ~     (p=0.310 n=5+5)
Chain/chain-1/Explore                     15.1µs ± 3%  15.1µs ± 2%     ~     (p=0.548 n=5+5)
Chain/chain-2/Explore                     87.7µs ± 1%  89.0µs ± 2%     ~     (p=0.151 n=5+5)
Chain/chain-3/Explore                      170µs ± 6%   175µs ± 8%     ~     (p=0.421 n=5+5)
Chain/chain-7/Explore                      528µs ± 1%   528µs ± 1%     ~     (p=1.000 n=5+5)
Chain/chain-8/Explore                      635µs ± 1%   634µs ± 1%     ~     (p=0.841 n=5+5)
Chain/chain-9/Explore                      792µs ± 4%   788µs ± 1%     ~     (p=1.000 n=5+5)
Chain/chain-11/Explore                    1.12ms ± 5%  1.09ms ± 2%     ~     (p=0.222 n=5+5)
Chain/chain-13/Explore                    1.40ms ± 3%  1.46ms ± 3%     ~     (p=0.056 n=5+5)
Chain/chain-14/Explore                    1.64ms ± 2%  1.67ms ± 3%     ~     (p=0.222 n=5+5)
Chain/chain-15/Explore                    1.82ms ± 1%  1.86ms ± 2%     ~     (p=0.056 n=5+5)
Chain/chain-16/Explore                    2.04ms ± 3%  2.05ms ± 1%     ~     (p=0.421 n=5+5)
Chain/chain-18/Explore                    2.54ms ± 3%  2.55ms ± 2%     ~     (p=1.000 n=5+5)
Chain/chain-19/Explore                    2.77ms ± 1%  2.80ms ± 1%     ~     (p=0.056 n=5+5)
FKInsert/SingleRow/None                    647µs ± 2%   649µs ± 2%     ~     (p=0.548 n=5+5)
FKInsert/SingleRow/NoFastPath              770µs ± 2%   799µs ± 5%     ~     (p=0.222 n=5+5)
FKInsert/SingleRow/FastPath                380µs ± 2%   390µs ± 4%     ~     (p=0.222 n=5+5)
FKInsert/MultiRowSingleParent/None        1.38ms ± 3%  1.41ms ± 5%     ~     (p=0.310 n=5+5)
FKInsert/MultiRowSingleParent/NoFastPath  1.28ms ± 5%  1.31ms ± 7%     ~     (p=0.548 n=5+5)
FKInsert/MultiRowSingleParent/FastPath     615µs ± 3%   630µs ± 5%     ~     (p=0.151 n=5+5)
FKInsert/MultiRowMultiParent/NoFastPath   1.40ms ± 4%  1.45ms ± 6%     ~     (p=0.151 n=5+5)
FKInsert/MultiRowMultiParent/FastPath      730µs ± 7%   738µs ± 2%     ~     (p=0.310 n=5+5)
Chain/chain-4/Explore                      245µs ± 1%   247µs ± 1%   +0.95%  (p=0.032 n=5+5)
Chain/chain-6/Explore                      423µs ± 1%   429µs ± 2%   +1.39%  (p=0.032 n=5+5)
Chain/chain-17/Explore                    2.26ms ± 2%  2.31ms ± 1%   +2.10%  (p=0.016 n=5+5)
Phases/tpcc-new-order/ExecBuild           43.9µs ± 2%  45.1µs ± 3%   +2.63%  (p=0.032 n=5+5)
Phases/kv-read/Parse                      2.53ns ± 2%  2.59ns ± 1%   +2.69%  (p=0.016 n=5+5)
Phases/kv-read-const/Normalize             150ns ± 3%   156ns ± 0%   +3.83%  (p=0.016 n=5+4)
Phases/kv-read-const/OptBuild              151ns ± 2%   157ns ± 2%   +3.96%  (p=0.008 n=5+5)
Chain/chain-12/Explore                    1.21ms ± 1%  1.27ms ± 2%   +4.32%  (p=0.008 n=5+5)
FKInsert/MultiRowMultiParent/None         1.48ms ± 2%  1.57ms ± 2%   +6.24%  (p=0.008 n=5+5)
Phases/tpcc-delivery/OptBuild             15.0µs ± 6%  16.0µs ± 5%   +6.49%  (p=0.032 n=5+5)
Phases/kv-read-no-prep/Explore            50.8µs ± 7%  55.7µs ± 9%   +9.69%  (p=0.032 n=5+5)
```